requirements.txt
torch>=2.0.0
transformers>=4.37.0
datasets>=2.14.0
peft>=0.5.0
bitsandbytes>=0.41.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-readers-file>=0.1.0
python-docx>=0.8.11
pandas>=2.0.0
sentence-transformers>=2.2.2
accelerate>=0.25.0
scikit-learn>=1.3.0
tqdm>=4.66.0
numpy>=1.24.0 